+++ src/main/java/com/agroproserver/serveragropro/dto/request/ParcelaRequestDto.java
@@ -1,5 +1,6 @@
 package com.agroproserver.serveragropro.dto.request;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 import jakarta.validation.constraints.NotBlank;
@@ -25,7 +26,7 @@ public class ParcelaRequestDto {
     private String usoPrincipal;
 
     @NotNull
-    private Long superficie;
+    private BigDecimal superficie;
 
     private Long valorSuelo;
